# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QuantPilot Operator Pre-Harness",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/policies/parse": {
      "post": {
        "summary": "Parse Policy",
        "operationId": "parse_policy_api_policies_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParsePolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPolicy"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/autopilot/guarded/run-once": {
      "post": {
        "summary": "Guarded Run Once",
        "operationId": "guarded_run_once_api_autopilot_guarded_run_once_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutopilotPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Guarded Run Once Api Autopilot Guarded Run Once Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/autopilot/guarded/pause": {
      "post": {
        "summary": "Pause Guarded",
        "operationId": "pause_guarded_api_autopilot_guarded_pause_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutopilotPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Pause Guarded Api Autopilot Guarded Pause Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/autopilot/guarded/resume": {
      "post": {
        "summary": "Resume Guarded",
        "operationId": "resume_guarded_api_autopilot_guarded_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutopilotPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Resume Guarded Api Autopilot Guarded Resume Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/autopilot/kill-switch": {
      "post": {
        "summary": "Kill Switch",
        "operationId": "kill_switch_api_autopilot_kill_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Kill Switch Api Autopilot Kill Switch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/autopilot/kill-switch/release": {
      "post": {
        "summary": "Release Kill Switch",
        "operationId": "release_kill_switch_api_autopilot_kill_switch_release_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Release Kill Switch Api Autopilot Kill Switch Release Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/operator/run-once": {
      "post": {
        "summary": "Operator Run Once",
        "operationId": "operator_run_once_api_operator_run_once_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorRunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AutopilotPolicyRequest": {
        "properties": {
          "policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Id"
          }
        },
        "type": "object",
        "title": "AutopilotPolicyRequest"
      },
      "BrokerMode": {
        "type": "string",
        "enum": [
          "mock",
          "paper",
          "live_disabled"
        ],
        "title": "BrokerMode"
      },
...
      "ExecutionMode": {
        "type": "string",
        "enum": [
          "backtest_only",
          "paper_trading",
          "approval_required",
          "guarded_autopilot",
          "fully_automated"
        ],
        "title": "ExecutionMode"
      },
      "FallbackDecision": {
        "properties": {
          "from_level": {
            "type": "integer",
            "const": 5,
            "title": "From Level"
          },
          "to_level": {
            "type": "integer",
            "enum": [
              4,
              3,
              2,
              0
            ],
            "title": "To Level"
          },
          "reason_code": {
            "type": "string",
            "title": "Reason Code"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "order_submission_enabled": {
            "type": "boolean",
            "title": "Order Submission Enabled",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "from_level",
          "to_level",
          "reason_code",
          "detail"
        ],
        "title": "FallbackDecision"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KillSwitchRequest": {
        "properties": {
          "policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": "user_requested"
          }
        },
        "type": "object",
        "title": "KillSwitchRequest"
      },
...
      "OperatorDecision": {
        "properties": {
          "decision_id": {
            "type": "string",
            "title": "Decision Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "policy_version": {
            "type": "integer",
            "title": "Policy Version"
          },
          "strategy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Id"
          },
          "order_plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Plan Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "submit",
              "block",
              "fallback",
              "noop"
            ],
            "title": "Action"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "risk_check_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Check Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "policy_id",
          "policy_version",
          "action",
          "reason"
        ],
        "title": "OperatorDecision"
      },
      "OperatorReport": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "policy_version": {
            "type": "integer",
            "title": "Policy Version"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "blocked",
              "fallback",
              "failed"
            ],
            "title": "Status"
          },
          "strategy_selection": {
            "$ref": "#/components/schemas/StrategySelectionDecision"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/OperatorDecision"
            },
            "type": "array",
            "title": "Decisions"
          },
          "fallback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FallbackDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "order_plan_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Order Plan Ids"
          },
          "broker_order_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Broker Order Ids"
          },
          "risk_check_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Check Ids"
          },
          "safety_flags": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "object",
            "title": "Safety Flags"
          },
          "live_trading_enabled": {
            "type": "boolean",
            "title": "Live Trading Enabled",
            "default": false
          },
          "audit_event_count": {
            "type": "integer",
            "title": "Audit Event Count",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "user_id",
          "policy_id",
          "policy_version",
          "started_at",
          "completed_at",
          "status",
          "strategy_selection",
          "decisions"
        ],
        "title": "OperatorReport"
      },
      "OperatorRunRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "fixture-user"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "requested_policy_version": {
            "type": "integer",
            "title": "Requested Policy Version"
          },
          "run_mode": {
            "type": "string",
            "enum": [
              "dry_run",
              "mock_submit",
              "paper_submit"
            ],
            "title": "Run Mode",
            "default": "dry_run"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "policy_id",
          "requested_policy_version",
          "idempotency_key"
        ],
        "title": "OperatorRunRequest"
      },
      "OperatorRunResult": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "blocked",
              "fallback",
              "failed"
            ],
            "title": "Status"
          },
          "submitted_order_plan_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Submitted Order Plan Ids"
          },
          "blocked_order_plan_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocked Order Plan Ids"
          },
          "fallback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FallbackDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "report": {
            "$ref": "#/components/schemas/OperatorReport"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "status",
          "report"
        ],
        "title": "OperatorRunResult"
      },
...
      "OrderType": {
        "type": "string",
        "enum": [
          "limit",
          "market"
        ],
        "title": "OrderType"
      },
      "ParsePolicyRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "default": "KR stock moderate risk weekly rebalance, approval required, mock broker, limit orders only."
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "fixture-user"
          }
        },
        "type": "object",
        "title": "ParsePolicyRequest"
      },
...
      "ReleaseKillSwitchRequest": {
        "properties": {
          "policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Id"
          },
          "confirmation": {
            "type": "string",
            "title": "Confirmation"
          }
        },
        "type": "object",
        "required": [
          "confirmation"
        ],
        "title": "ReleaseKillSwitchRequest"
      },
...
      "StrategySelectionDecision": {
        "properties": {
          "selected_strategy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Strategy Id"
          },
          "selected_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Version"
          },
          "eligible_strategy_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Eligible Strategy Ids"
          },
          "rejected": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Rejected"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "selected_strategy_id",
          "selected_version",
          "eligible_strategy_ids",
          "rejected",
          "reason"
        ],
        "title": "StrategySelectionDecision"
      },
      "UserPolicy": {
        "properties": {
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "fixture-user"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "market": {
            "type": "string",
            "title": "Market",
            "default": "KR_STOCK"
          },
          "risk_profile": {
            "type": "string",
            "title": "Risk Profile",
            "default": "moderate"
          },
          "max_positions": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Max Positions",
            "default": 8
          },
          "max_position_weight": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Position Weight",
            "default": 0.15
          },
          "max_sector_weight": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Sector Weight",
            "default": 0.4
          },
          "min_cash_weight": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "Min Cash Weight",
            "default": 0.2
          },
          "daily_loss_limit": {
            "type": "number",
            "title": "Daily Loss Limit",
            "default": -0.03
          },
          "monthly_loss_limit": {
            "type": "number",
            "title": "Monthly Loss Limit",
            "default": -0.05
          },
          "max_daily_orders": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Max Daily Orders",
            "default": 3
          },
          "max_daily_turnover": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Daily Turnover",
            "default": 3000000
          },
          "monthly_loss_pause_new_buys": {
            "type": "number",
            "title": "Monthly Loss Pause New Buys",
            "default": -0.05
          },
          "monthly_loss_stop_all_autotrading": {
            "type": "number",
            "title": "Monthly Loss Stop All Autotrading",
            "default": -0.1
          },
          "stale_quote_max_age_seconds": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Stale Quote Max Age Seconds",
            "default": 30
          },
          "human_review_quote_max_age_seconds": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Human Review Quote Max Age Seconds",
            "default": 120
          },
          "order_expiry_minutes": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Order Expiry Minutes",
            "default": 30
          },
          "authority_level": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Authority Level",
            "default": 2
          },
          "kill_switch_engaged": {
            "type": "boolean",
            "title": "Kill Switch Engaged",
            "default": false
          },
          "guarded_autopilot_enabled": {
            "type": "boolean",
            "title": "Guarded Autopilot Enabled",
            "default": false
          },
          "fully_automated_operator_enabled": {
            "type": "boolean",
            "title": "Fully Automated Operator Enabled",
            "default": false
          },
          "single_order_cash_limit": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Single Order Cash Limit",
            "default": 1000000
          },
          "rebalance_frequency": {
            "type": "string",
            "title": "Rebalance Frequency",
            "default": "weekly"
          },
          "execution_mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "default": "approval_required"
          },
          "allowed_order_types": {
            "items": {
              "$ref": "#/components/schemas/OrderType"
            },
            "type": "array",
            "title": "Allowed Order Types"
          },
          "broker": {
            "$ref": "#/components/schemas/BrokerMode",
            "default": "mock"
          },
          "preferred_themes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Themes"
          },
          "preferred_sectors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Sectors"
          },
          "blocklist": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocklist"
          },
          "min_avg_daily_value": {
            "type": "number",
            "minimum": 0.0,
            "title": "Min Avg Daily Value",
            "default": 5000000
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UserPolicy"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}